moved autocompleter class and it's related helper classes and enums into the autocompleter package
