+++ src/main/java/seedu/address/ui/autocompleter/AutocompleteState.java
@@ -1,4 +1,4 @@
-package seedu.address.ui;
+package seedu.address.ui.autocompleter;
 
 //@@author Kowalski985
 /**
